+++ scripts/lint-samples.js
@@ -1,0 +1,45 @@
+const { execSync } = require('child_process');
+const samples = require('./samples.json');
+const { getAppName } = require('./utils');
+
+/**
+ * Start linting process only for the samples that were affected by the new changes in:
+ * - create-sitecore-jss/src/templates/**
+ * - create-sitecore-jss/src/initializers/**
+ */
+
+const affectedTemplates = execSync(
+  'git diff --name-only origin/dev... -- packages/create-sitecore-jss',
+  {
+    encoding: 'utf-8',
+  }
+)
+  .split('\n')
+  .map((filepath) => {
+    // Extracting template names from the filepath
+    const template = filepath.match(/(templates|initializers)\/([^\/]*)/);
+
+    return template && template[2];
+  })
+  // Removing null values and leaving unique template names
+  .filter((template, index, list) => template && list.indexOf(template) === index);
+
+if (!affectedTemplates.length) {
+  console.log('No modified templates to lint');
+
+  return;
+}
+
+const affectedSamples = samples
+  .filter((sample) => sample.initializers.some((template) => affectedTemplates.includes(template)))
+  .map((sample) => getAppName(sample.args));
+
+if (!affectedSamples.length) {
+  console.log('No modified samples to lint');
+
+  return;
+}
+
+execSync(`npx lerna run lint --scope={${affectedSamples.join(',')}} -- --fix`, {
+  stdio: 'inherit',
+});